refactor(conversation): extract the queued mutation controller base
