+++ src/aiSessions/conversation/queuedCommentController.ts
@@ -1,50 +1,34 @@
 'use strict';
 
-import * as vscode from 'vscode';
-import type { AiSessionProviderId } from '../../models';
 import type { CommentErrorCode, CommentStatus } from './commentPrimitives';
 import { CommentError } from './commentPrimitives';
+import {
+    QueuedMutationController,
+    QueuedMutationControllerOptions,
+    QueuedMutationRequestBase,
+    QueuedMutationResultBase,
+} from './queuedMutationController';
 import type { CommentSnapshot } from './snapshotFileStore';
 import type { ConversationViewerTarget } from './viewerTarget';
 
 /**
- * Shared resilience protocol for the comment-stack controllers: serialized
- * operations, optimistic revisions, persist → commit → rollback, idempotent
- * settlements with replay, and a rebuild fallback when the webview goes
- * away. Each stack supplies its model operations through the abstract
+ * Comment-stack specialization of the queued mutation pipeline: adds the
+ * mutate/send choreography (prompt building, done-marking or dispatch
+ * recording, submit rollback, focus-after-send) on top of the shared
+ * serialized operations, optimistic revisions, and idempotent settlements.
+ * Each comment stack supplies its model operations through the abstract
  * hooks; everything else lives here exactly once.
  */
 
-export interface QueuedCommentRequestBase {
-    type: string;
-    version: 1;
-    requestId: string;
-    subscriptionGeneration: number;
-    projectId: string;
-    provider: AiSessionProviderId;
-    sessionId: string;
-    operation: string;
-    expectedRevision: number;
-    payload: unknown;
-}
+export interface QueuedCommentRequestBase extends QueuedMutationRequestBase {}
 
-export interface QueuedCommentResultBase<TComment> {
-    type: string;
-    version: 1;
-    requestId: string;
-    subscriptionGeneration: number;
-    projectId: string;
-    provider: AiSessionProviderId;
-    sessionId: string;
-    operation: string;
-    success: boolean;
-    revision: number;
+export interface QueuedCommentResultBase<TComment>
+    extends QueuedMutationResultBase {
     comments: TComment[];
-    error?: CommentErrorCode;
 }
 
-export interface QueuedCommentControllerOptions {
-    now?: () => number;
+export interface QueuedCommentControllerOptions
+    extends QueuedMutationControllerOptions {
     submitPrompt: (
         target: ConversationViewerTarget,
         prompt: string
@@ -55,18 +39,6 @@ export interface QueuedCommentControllerOptions {
             'projectId' | 'provider' | 'sessionId'
         >
     ) => PromiseLike<void> | Promise<void>;
-    getTarget: () => ConversationViewerTarget | undefined;
-    getSubscriptionGeneration: () => number;
-    getPanel: () => vscode.WebviewPanel | undefined;
-    rebuildLatestDocument: () => void;
-}
-
-interface CommentSnapshotStore<TStoreTarget, TComment> {
-    load(target: TStoreTarget): Promise<CommentSnapshot<TComment>>;
-    save(
-        target: TStoreTarget,
-        snapshot: CommentSnapshot<TComment>
-    ): Promise<void>;
 }
 
 export abstract class QueuedCommentController<
@@ -74,133 +46,73 @@ export abstract class QueuedCommentController<
     TStoreTarget,
     TRequest extends QueuedCommentRequestBase,
     TResult extends QueuedCommentResultBase<TComment>
+> extends QueuedMutationController<
+    TStoreTarget,
+    CommentSnapshot<TComment>,
+    TRequest,
+    TResult
 > {
     private comments: TComment[] = [];
-    private revision = 0;
-    private operationQueue: Promise<void> = Promise.resolve();
-    private readonly settlements = new Map<string, TResult>();
 
     constructor(
         protected readonly options: QueuedCommentControllerOptions
-    ) {}
-
-    protected now(): number {
-        return this.options.now?.() ?? Date.now();
+    ) {
+        super(options);
     }
 
-    get snapshot(): CommentSnapshot<TComment> {
-        return {
-            revision: this.revision,
-            comments: this.cloneComments(this.comments),
-        };
+    /** Read-only view for subclass mutation branches. */
+    protected get currentComments(): readonly TComment[] {
+        return this.comments;
     }
 
-    reset(): void {
+    protected clearState(): void {
         this.comments = [];
-        this.revision = 0;
-        this.settlements.clear();
-        this.onReset();
     }
 
-    async drainMutations(): Promise<void> {
-        await this.operationQueue;
+    protected statePayload(): object {
+        return { comments: this.cloneComments(this.comments) };
     }
 
-    enqueue(request: TRequest): Promise<void> {
-        return this.enqueueTask(() => this.handleOperation(request));
+    protected validateRestoredSnapshot(
+        snapshot: CommentSnapshot<TComment>
+    ): boolean {
+        try {
+            this.validateComments(snapshot.comments);
+            return Number.isSafeInteger(snapshot.revision)
+                && snapshot.revision >= 0;
+        } catch (_error) {
+            return false;
+        }
     }
 
-    /** Serializes an arbitrary task behind the queued mutations. */
-    protected enqueueTask(operation: () => Promise<void>): Promise<void> {
-        const queued = this.operationQueue.then(operation, operation);
-        this.operationQueue = queued.catch(() => undefined);
-        return queued;
+    protected applyRestoredSnapshot(
+        snapshot: CommentSnapshot<TComment>
+    ): void {
+        this.comments = this.cloneComments(snapshot.comments);
     }
 
-    async restore(
+    protected async performRequest(
+        request: TRequest,
         target: ConversationViewerTarget,
         generation: number
     ): Promise<void> {
-        const store = this.getStore();
-        if (!store) {
+        if (this.isSendRequest(request)) {
+            await this.sendComments(
+                target,
+                generation,
+                this.sendCommentId(request)
+            );
             return;
         }
-        let snapshot: CommentSnapshot<TComment>;
-        try {
-            snapshot = await store.load(this.toStoreTarget(target));
-            this.validateComments(snapshot.comments);
-            if (!Number.isSafeInteger(snapshot.revision)
-                || snapshot.revision < 0) {
-                throw this.makeError('invalid');
-            }
-        } catch (_error) {
-            return;
-        }
-        if (this.options.getTarget() !== target
-            || this.options.getSubscriptionGeneration() !== generation) {
-            return;
-        }
-        this.comments = this.cloneComments(snapshot.comments);
-        this.revision = snapshot.revision;
+        await this.mutateComments(request, target, generation);
     }
 
-    private async handleOperation(request: TRequest): Promise<void> {
-        const target = this.options.getTarget();
-        if (!target || !this.options.getPanel()) {
-            return;
-        }
-        const settlementKey = getSettlementKey(request);
-        const settled = this.settlements.get(settlementKey);
-        if (settled) {
-            await this.publishSettlement(
-                settled.operation === request.operation
-                    ? settled
-                    : {
-                        ...settled,
-                        operation: request.operation,
-                        success: false,
-                        revision: this.revision,
-                        comments: this.cloneComments(this.comments),
-                        error: 'invalid',
-                    } as TResult,
-                false
-            );
-            return;
-        }
-        if (this.isMutationsFrozen()
-            || !requestTargetsViewer(
-                request,
-                target,
-                this.options.getSubscriptionGeneration()
-            )
-            || request.expectedRevision !== this.revision) {
-            await this.settleRequest(request, false, 'stale');
-            return;
-        }
-        try {
-            if (this.isSendRequest(request)) {
-                await this.sendComments(
-                    target,
-                    this.options.getSubscriptionGeneration(),
-                    this.sendCommentId(request)
-                );
-            } else {
-                await this.mutateComments(
-                    request,
-                    target,
-                    this.options.getSubscriptionGeneration()
-                );
-            }
-            await this.settleRequest(request, true);
-            if (this.isSendRequest(request)) {
-                await this.focusSessionAfterSend(target);
-            }
-        } catch (error) {
-            await this.settleRequest(
-                request,
-                false,
-                this.toErrorCode(error)
-            );
+    protected async afterSuccessful(
+        request: TRequest,
+        target: ConversationViewerTarget
+    ): Promise<void> {
+        if (this.isSendRequest(request)) {
+            await this.focusSessionAfterSend(target);
         }
     }
 
@@ -224,26 +136,30 @@ export abstract class QueuedCommentController<
         target: ConversationViewerTarget,
         generation: number
     ): Promise<void> {
-        const { comments, changed } = this.applyMutation(request);
-        if (!changed) {
+        const mutation = this.applyMutation(request);
+        if (!mutation.changed) {
             // A no-op mutation (e.g. adding a duplicate tag) still settles
             // successfully, just without touching the persisted snapshot.
             return;
         }
-        const snapshot = { revision: this.revision + 1, comments };
+        const snapshot = {
+            revision: this.revision + 1,
+            comments: mutation.comments,
+        };
         const previousSnapshot = this.snapshot;
         await this.persist(target, snapshot);
         try {
-            this.commitPersisted(
+            this.assertUnchanged(
                 target,
                 generation,
-                request.expectedRevision,
-                snapshot
+                request.expectedRevision
             );
         } catch (error) {
             await this.persist(target, previousSnapshot);
             throw error;
         }
+        this.comments = snapshot.comments;
+        this.revision = snapshot.revision;
     }
 
     private async sendComments(
@@ -268,11 +184,15 @@ export abstract class QueuedCommentController<
             comments: this.buildSendComments(targetComments, target),
         };
         await this.persist(target, sentSnapshot);
-        if (this.options.getTarget() !== target
-            || this.options.getSubscriptionGeneration() !== generation
-            || this.revision !== previousSnapshot.revision) {
+        try {
+            this.assertUnchanged(
+                target,
+                generation,
+                previousSnapshot.revision
+            );
+        } catch (error) {
             await this.persist(target, previousSnapshot);
-            throw this.makeError('stale');
+            throw error;
         }
         try {
             await Promise.resolve(this.options.submitPrompt(
@@ -284,9 +204,7 @@ export abstract class QueuedCommentController<
             // submitPrompt surfaces submission.ts failures; both stacks'
             // error classes share the CommentError base, so user-actionable
             // codes survive the cross-stack hop.
-            throw error instanceof CommentError
-                ? this.makeError(error.code)
-                : this.makeError('failed');
+            throw this.makeError(this.toErrorCode(error));
         }
         if (this.options.getTarget() !== target) {
             throw this.makeError('stale');
@@ -295,119 +213,6 @@ export abstract class QueuedCommentController<
         this.revision = sentSnapshot.revision;
     }
 
-    private async persist(
-        target: ConversationViewerTarget,
-        snapshot: CommentSnapshot<TComment>
-    ): Promise<void> {
-        const store = this.getStore();
-        if (!store) {
-            return;
-        }
-        try {
-            await store.save(this.toStoreTarget(target), snapshot);
-        } catch (_error) {
-            throw this.makeError('failed');
-        }
-    }
-
-    private commitPersisted(
-        target: ConversationViewerTarget,
-        generation: number,
-        expectedRevision: number,
-        snapshot: CommentSnapshot<TComment>
-    ): void {
-        if (this.options.getTarget() !== target
-            || this.options.getSubscriptionGeneration() !== generation
-            || this.revision !== expectedRevision) {
-            throw this.makeError('stale');
-        }
-        this.comments = snapshot.comments;
-        this.revision = snapshot.revision;
-    }
-
-    private async settleRequest(
-        request: TRequest,
-        success: boolean,
-        error?: CommentErrorCode
-    ): Promise<void> {
-        const settlement = {
-            type: this.resultType,
-            version: 1,
-            requestId: request.requestId,
-            subscriptionGeneration: request.subscriptionGeneration,
-            projectId: request.projectId,
-            provider: request.provider,
-            sessionId: request.sessionId,
-            operation: request.operation,
-            success,
-            revision: this.revision,
-            comments: this.cloneComments(this.comments),
-            ...(error ? { error } : {}),
-        } as TResult;
-        this.rememberSettlement(getSettlementKey(request), settlement);
-        await this.publishSettlement(settlement, true);
-    }
-
-    private rememberSettlement(key: string, settlement: TResult): void {
-        this.settlements.set(key, settlement);
-        while (this.settlements.size > 100) {
-            const oldest = this.settlements.keys().next().value;
-            if (typeof oldest !== 'string') {
-                break;
-            }
-            this.settlements.delete(oldest);
-        }
-    }
-
-    protected async publishSettlement(
-        settlement: object,
-        rebuildOnFailure: boolean
-    ): Promise<void> {
-        const panel = this.options.getPanel();
-        if (!panel) {
-            return;
-        }
-        let delivered = false;
-        try {
-            delivered = await panel.webview.postMessage(settlement);
-        } catch (_error) {
-            delivered = false;
-        }
-        if (!delivered
-            && rebuildOnFailure
-            && this.options.getPanel() === panel) {
-            this.options.rebuildLatestDocument();
-        }
-    }
-
-    protected toErrorCode(error: unknown): CommentErrorCode {
-        return error instanceof CommentError
-            ? error.code
-            : 'failed';
-    }
-
-    /** Read-only view for subclass mutation branches. */
-    protected get currentComments(): readonly TComment[] {
-        return this.comments;
-    }
-
-    /** Hook for subclass state cleared alongside the base state. */
-    protected onReset(): void {}
-
-    /** Frozen stacks reject mutations as 'stale' (session rebind windows). */
-    protected isMutationsFrozen(): boolean {
-        return false;
-    }
-
-    protected abstract readonly resultType: TResult['type'];
-
-    protected abstract getStore():
-        CommentSnapshotStore<TStoreTarget, TComment> | undefined;
-
-    protected abstract toStoreTarget(
-        target: ConversationViewerTarget
-    ): TStoreTarget;
-
     protected abstract cloneComments(
         comments: readonly TComment[]
     ): TComment[];
@@ -415,8 +220,6 @@ export abstract class QueuedCommentController<
     protected abstract validateComments(
         comments: readonly TComment[]
     ): void;
-
-    protected abstract makeError(code: CommentErrorCode): CommentError;
 
     protected abstract isSendRequest(request: TRequest): boolean;
 
@@ -441,24 +244,4 @@ export abstract class QueuedCommentController<
         targetComments: readonly TComment[],
         target: ConversationViewerTarget
     ): TComment[];
-}
-
-function requestTargetsViewer(
-    request: QueuedCommentRequestBase,
-    target: ConversationViewerTarget,
-    subscriptionGeneration: number
-): boolean {
-    return request.subscriptionGeneration === subscriptionGeneration
-        && request.projectId === target.projectId
-        && request.provider === target.provider
-        && request.sessionId === target.sessionId;
-}
-
-function getSettlementKey(request: QueuedCommentRequestBase): string {
-    return JSON.stringify([
-        request.projectId,
-        request.provider,
-        request.sessionId,
-        request.requestId,
-    ]);
 }
